+++ packages/connect-iframe/src/index.ts
@@ -29,7 +29,7 @@ import {
     suggestBridgeInstaller,
     suggestUdevInstaller,
 } from '@trezor/connect/src/utils/browserUtils';
-import * as storage from '@trezor/connect-common/src/storage';
+import { storage } from '@trezor/connect-common';
 
 let _core: Core | undefined;
 
@@ -185,8 +185,7 @@ const filterDeviceEvent = (message: DeviceEvent) => {
     const features =
         'device' in message.payload ? message.payload.device.features : message.payload.features;
     if (features) {
-        const savedPermissions =
-            storage.load(storage.PERMISSIONS_KEY) || storage.load(storage.PERMISSIONS_KEY, true);
+        const savedPermissions = storage.load().permissions || storage.load(true).permissions;
         if (savedPermissions) {
             const devicePermissions = savedPermissions.filter(
                 p =>